清理代码

+++ particleSystem.cuh
@@ -4,7 +4,7 @@ extern "C"
 {
     void cudaInit(int argc, char **argv);
 
-    void allocateArray(void **devPtr, int size);
+    void allocateArray(void **devPtr, size_t size);
     void freeArray(void *devPtr);
 
     void threadSync();
